requirements.txt
portal
gymnasium
click
numpy
pyyaml
wget
loguru